requirements.txt
pillow
realesrgan
torch
torchvision==0.13.1
numpy<2
opencv-contrib-python
pypdf